splitted the library and generated shared object
